requirements.txt
agent-framework-core>=1.0.1
agent-framework-openai>=1.0.1
python-dotenv>=1.0.1
httpx>=0.27.0
pytest>=8.2.0
pyobjc-framework-EventKit>=10.0
